pyproject.toml
[tool.poetry]
name = "ttrack"
version = "0.2.0"
description = ""
authors = ["Alexander Astashov <alexander.astashov.vl@gmail.com>"]
readme = "README.md"
packages = [{include = "ttrack", from = "src"}]

[tool.poetry.scripts]
ttrack = "ttrack.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
workdays = "^1.4"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
coloredlogs = "^15.0.1"
pyperclip = "^1.8.2"
python-dateutil = "^2.8.2"
aiohttp = "^3.8.3"
cleo = "^2.0.1"

[tool.poetry.group.local.dependencies]
ipython = "^8.7.0"

[tool.poetry.group.qa.dependencies]
pycln = "^2.1.2"
black = "^22.10.0"
isort = "^5.10.1"
pyproject-flake8 = "^6.0.0"
types-requests = "^2.28.11.5"
types-python-dateutil = "^2.8.19.4"

[tool.black]
line-length = 120
exclude = '\.venv'

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files.html
profile = "black"
line_length = 120
skip = [".cache"]

[tool.flake8]
max-line-length = 120

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
